+++ CHANGELOG.md
@@ -7,6 +7,7 @@ HumHub Changelog
 - Fix #7720: Saving the new Mobile App admin form when no whitelisted domain
 - Fix: Bootstrap 3 compatible SCSS overwrites unwanted values such as Markdown h1 font size
 - Fix #7722: Fix mailer settings loss after upgrade
+- Fix #7736: Fix scroll of meta search modal window
 
 1.18.0-beta.4 (September 19, 2025)
 ----------------------------------